src/test/java/com/ada/ecosystem/core/v1/database/EcosystemInterceptorStatementTest.java
package com.ada.ecosystem.core.v1.database;

import static org.junit.jupiter.api.Assertions.assertEquals;

import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;

import com.ada.ecosystem.core.v1.utility.EcosystemUtilities;

import jakarta.persistence.EntityManager;

public class EcosystemInterceptorStatementTest {

    @Mock
    private EntityManager em;

    @InjectMocks
    private EcosystemInterceptorStatement interceptor;

    @BeforeEach
    public void setUp() {
        MockitoAnnotations.openMocks(this);
        // No static mocking here as it's not allowed in Mockito without PowerMockito
    }


    @Test
    public void testInspectWithCodigoMempresaSql() {
        // Arrange
        String sql = "codigo_mempresa";
        EcosystemUtilities.codigoMempresa = "testCode"; // Set static field for test

        // Act
        String result = interceptor.inspect(sql);

        // Assert
        assertEquals(sql, result);
        // Note: No direct way to verify static methods without PowerMockito
    }

}
